labels/table: Fix bugs on 32 bit platforms
The cast a long type is probably a remnant of gcli_id being a typedef
for unsigned long. In this case we observed issues when passing an
ID and this cast somehow fixed the bug - I don't recall all the
details but I remember fixing this very vividly on a machine with
a VIA Eden C7 Processor running FreeBSD :^)

Fix this by using the right types everywhere.

Reported-by: Simon Richter <Simon.Richter@hogyros.de>
References: https://todo.sr.ht/~herrhotzenplotz/gcli/10
References: aa4bda208f97d86b8f38b5aaf652af8880125543 (Make gcli_id type a uint64_t)

+++ src/cmd/labels.c
@@ -92,7 +92,7 @@ gcli_labels_print(struct gcli_label_list const *const list, int const max)
 
 	for (size_t i = 0; i < n; ++i) {
 		gcli_tbl_add_row(table,
-		                 (long)list->labels[i].id, /* Cast is important here (#165) */
+		                 list->labels[i].id,
 		                 list->labels[i].colour,
 		                 gcli_config_have_colours(g_clictx) ? "  " : "",
 		                 list->labels[i].name,
